refactor: update imports in repositories, migrations and tests
- Update Base import in repositories to use backend.models.core
- Update model imports in equipment repository to use specific modules
- Update imports in Alembic migration environment
- Update test configuration to use new import structure
